Add single producer single consume thread tests
We try to keep push_count and pop_count pinned to push/pop methods, ie.,
only push_count/pop_count will be modified push/pop methods respectively.

Yes, there is a possibility that these methods will return full/empty
errors even when these buffers are not; but that's okay since the data
in the buffer is not being corrupted; a subsequent push/pop will see
correct results. This is needed to keep a single producer and single
consumer usage pattern lock-free.

Add tests that make sure our expectations are met in this scenario.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>